add logout (complited logout method)

+++ server/src/services/auth/auth.controller.ts
@@ -38,9 +38,10 @@ export class AuthController {
 
     @Get('logout')
     async logout(@Cookie(REFRESH_TOKEN) refreshToken: string, @Res() res: Response) {
+        if (!refreshToken) return res.sendStatus(HttpStatus.OK)
         await this.authService.deleteRefreshTokeb(refreshToken)
-        res.cookie(REFRESH_TOKEN, '', { httpOnly: true, expires: new Date() })
-        res.status(HttpStatus.OK)
+        res.cookie(REFRESH_TOKEN, '', { httpOnly: true, secure: true, expires: new Date() })
+        res.sendStatus(HttpStatus.OK)
     }
 
     // ------------------------------------------ Tokens ------------------------------------------ //
